+++ src/components/employee-beta/cards/BaseCard.vue
